fix(spirv): emit the fma instruction instead of expanding it
`math::FmaOp` lowered to `a * b + c`, so this backend never emitted
`OpExtInst GLSL.std.450 Fma`. It was the only one that did this: the C++
targets emit `fma`, `__hfma` or `__hfma2`, the LLVM target emits an FMA
intrinsic, and WGSL emits `fma(a, b, c)`. Anyone calling `fma()` for its
single rounding was quietly getting two roundings instead.

`gl::FmaOp` is defined for floats only, so an integer fma keeps its multiply
and add behind a `should_lower` predicate. That costs nothing: integer
arithmetic is exact, so there is no rounding to save, and instruction
selection still emits a multiply-add where the hardware has one.

The ternary mapper this needs already existed in `cmp.rs`, named after its
only user. It moves next to the unop and binop mappers, the three clamps use
it, and the duplicate goes away.
